Correction chapman motion blur values

+++ sources/engine/Stride.Rendering/Rendering/Images/MotionBlur/ChapmanBlurShader.sdsl.cs
@@ -14,7 +14,7 @@ namespace Stride.Rendering
 {
     public static partial class ChapmanBlurShaderKeys
     { 
-        public static readonly ValueParameterKey<uint> u_BlurRadius= ParameterKeys.NewValue<uint>();
+        public static readonly ValueParameterKey<float> u_BlurRadius= ParameterKeys.NewValue<float>();
         public static readonly ValueParameterKey<uint> u_MaxSamples = ParameterKeys.NewValue<uint>();
     }
 }
